+++ airbyte-cdk/python/source_declarative_manifest/main.py
@@ -11,7 +11,7 @@ from airbyte_cdk.connector import BaseConnector
 from airbyte_cdk.entrypoint import AirbyteEntrypoint, launch
 from airbyte_cdk.sources.declarative.declarative_source import DeclarativeSource
 from airbyte_cdk.sources.declarative.manifest_declarative_source import ManifestDeclarativeSource
-from connector_builder.connector_builder_handler import ConnectorBuilderHandler
+from connector_builder import connector_builder_handler
 
 
 def create_source(config: Mapping[str, Any]) -> DeclarativeSource:
@@ -42,15 +42,26 @@ def preparse(args: List[str]) -> Tuple[str, str]:
     return parsed.command, parsed.config
 
 
+def handle_connector_builder_request(source: DeclarativeSource, config: Mapping[str, Any]):
+    command = config.get("__command")
+    if command == "resolve_manifest":
+        return connector_builder_handler.resolve_manifest(source)
+    raise ValueError(f"Unrecognized command {command}.")
+
+
+def handle_connector_request(source: DeclarativeSource, args: List[str]):
+    # Verify that the correct args are present for the production codepaths.
+    AirbyteEntrypoint.parse_args(args)
+    launch(source, sys.argv[1:])
+
+
 def handle_request(args: List[str]):
     config = get_config_from_args(args)
     source = create_source(config)
     if "__command" in config:
-        ConnectorBuilderHandler(source).handle_request(config)
+        handle_connector_builder_request(source, config)
     else:
-        # Verify that the correct args are present for the production codepaths.
-        AirbyteEntrypoint.parse_args(args)
-        launch(source, sys.argv[1:])
+        handle_connector_request(source, args)
 
 
 if __name__ == "__main__":
